pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pr-review-agent"
version = "0.1.0"
description = "Production-ready automated GitHub PR review agent using LangGraph and Gemini"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # LangChain / LangGraph
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    # Google Gemini SDK
    "google-generativeai>=0.8.0",
    # GitHub API
    "PyGithub>=2.4.0",
    # Config & validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # HTTP
    "httpx>=0.27.0",
    # Retry
    "tenacity>=9.0.0",
    # Logging
    "structlog>=24.4.0",
    # Web server / webhook
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # CLI
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.21.0",        # httpx mocking
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
pr-review = "pr_review_agent.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/pr_review_agent"]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
    "N",   # pep8-naming
    "ANN", # flake8-annotations
]
ignore = [
    "ANN101", # missing type for self
    "ANN102", # missing type for cls
    "S101",   # assert allowed in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S", "ANN"]

# ── MyPy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["github.*", "structlog.*", "google.*", "google.generativeai.*"]
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/pr_review_agent --cov-report=term-missing --cov-report=html"
filterwarnings = ["error", "ignore::DeprecationWarning"]
